create placeholder files
