--- src/dataset.py
@@ -1,32 +1,0 @@
-import torch
-
-class AudioDataset(Dataset):
-    def __init__(self):
-        return
-
-    def __getitem__(self, idx):
-        return
-
-
-class TextDataset(Dataset):
-    def __init__(self):
-        return
-
-    def __getitem(self, idx):
-        return
-
-
-class VideoDataset(Dataset):
-    def __init__(self):
-        return
-
-    def __getitem(self, idx):
-        return
-
-
-class ImageDataset(Dataset):
-    def __init__(self):
-        return
-
-    def __getitem(self, idx):
-        return
