+++ src/text_description.py
@@ -1,0 +1,2 @@
+class TextDescription:
+    return
