+++ src/video_description.py
@@ -1,0 +1,7 @@
+class VideoDescription:
+    def __init__(self, samples, processor, model, model_id, model_kwargs=None,
+            verbose=False):
+        return
+
+    def generate_sample_descriptions(self, text):
+        return
